+++ CHANGELOG.md
@@ -5,6 +5,7 @@
 
 ### Bugs
 
+* Fix `$AWS_REGION` not set by `eval`/`exec` despite being documented #1277
 * Fix bugs parsing AccountId's with leading zero's on the command line #1366
 
 ## [v2.2.2] - 2026-05-17