+++ CHANGELOG.md
@@ -5,6 +5,7 @@
 
 ### Bugs
 
+* Fix `$AWS_REGION` not set by `eval`/`exec` despite being documented #1277
 * Fix bugs parsing AccountId's with leading zero's on the command line #1366
 * Default to `AuthWorkflow: device_code` when SSH or WSL sessions are detected #1371